+++ bin/cache-key.ts
@@ -1,1 +1,1 @@
-export default 'cf15fb62292a40df'
+export default '4fade398cfa541d1'